many small fixes but now the api is divided
api divided into smaller parts to have more of an overview of the
project. Our main file for the api is now in the head folder under
main.go. This file does the routing out to our other files in /api.
There we have devided them into user, stair, comment, picture and
error. But the error.go inly includes a struct to our HandleErrror
struct. This was the easiest way to include this struct into every
package