fixed with preview on photos and some other stuff
Shrinking the photos so that we can load them up faster on the website.
Both for stairs and for photos. The photos section now have a preview.
Then we can call to just show a preview of all the photos linked to a
stair. The preview function returns the preview and the picture id.
Witch results in that we later can call the photo_id to show a full
picture of the picture.